+++ src/app/model/power.ts
@@ -1,11 +1,11 @@
 export class Power{
-    private name: string = '';
-    private minimumAfinity: string = '';
-    private action: string = '';
-    private specialResistAction: string = '';
-    private effect: string = '';
-    private cost: string = '';
-    private duration: string = '';
+    private name = '';
+    private minimumAfinity = '';
+    private action = '';
+    private specialResistAction = '';
+    private effect = '';
+    private cost = '';
+    private duration = '';
 
     constructor(name: string){
         this.name = name;