+++ src/app/pages/info/info.page.ts
@@ -1,9 +1,16 @@
-import { Component } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
+import pkg from '../../../../package.json';
 
 @Component({
   selector: 'app-info',
   templateUrl: './info.page.html',
   styleUrls: ['./info.page.scss'],
 })
-export class InfoPage {
+export class InfoPage implements OnInit {
+
+  public version = '';
+
+  ngOnInit(): void {
+    this.version = pkg.version;
+  }
 }